Refactoring files.
