+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/home/HappyHourItem.kt
@@ -1,0 +1,65 @@
+package com.jordangellatly.hermosahappyhour.ui.home
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.jordangellatly.hermosahappyhour.model.Restaurant
+import com.jordangellatly.hermosahappyhour.model.sampleSearchRestaurantData
+import com.jordangellatly.hermosahappyhour.ui.components.HermosaHappyHourSurface
+import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
+
+@Composable
+fun HappyHourItem(
+    restaurant: Restaurant,
+    onRestaurantClick: (Long) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    HermosaHappyHourSurface(
+        shape = MaterialTheme.shapes.medium,
+        modifier = modifier.padding(
+            start = 4.dp,
+            end = 4.dp,
+            bottom = 8.dp
+        )
+    ) {
+        Column(
+            horizontalAlignment = Alignment.CenterHorizontally,
+            modifier = Modifier
+                .clickable(onClick = { onRestaurantClick(restaurant.id) })
+                .padding(8.dp)
+        ) {
+            RestaurantImage(
+                imageUrl = restaurant.companyLogoUrl,
+                elevation = 4.dp,
+                contentDescription = null,
+                modifier = Modifier.size(120.dp)
+            )
+            Text(
+                text = restaurant.name,
+                style = MaterialTheme.typography.subtitle1,
+                color = HermosaHappyHourTheme.colors.textSecondary,
+                modifier = Modifier.padding(top = 8.dp)
+            )
+        }
+    }
+}
+
+@Preview
+@Composable
+fun HappyHourItemPreview() {
+    HermosaHappyHourTheme {
+        val restaurant = sampleSearchRestaurantData.first()
+        HappyHourItem(
+            restaurant = restaurant,
+            onRestaurantClick = {}
+        )
+    }
+}
